+++ lib/handlers/index.js
@@ -1,5 +1,4 @@
-const db = require("../../config/connection")
-const cTable = require("console.table")
+const db = require("../../db/connection")
 
 const viewHandlers = require("./view")
 const addHandlers = require("./add")
@@ -39,6 +38,7 @@ const execute = async (sql, message) => {
 const textOnly = /[a-zA-Z]+[\s\w]+/
 
 const handlers = {
+  end: () => db.end(),
   view: viewHandlers(display),
   add: addHandlers(select, textOnly, execute),
   update: updateHandlers(select, execute),
